fix: убрал выделение блока смены пароля в отдельный компонент и реализовал его непосредственно на странице профиля
