+++ packages/client/src/pages/ProfilePage/ProfilePage.module.scss
@@ -1,3 +1,5 @@
+$form-max-width: 510px;
+
 .profile {
   position: relative;
 
@@ -97,7 +99,7 @@
   }
 
   &__form {
-    max-width: 510px;
+    max-width: $form-max-width;
     margin: 3rem auto 0;
 
     label {
@@ -117,5 +119,25 @@
         color: currentcolor;
       }
     }
+  }
+}
+
+.modalForm {
+  &__container {
+    width: 100%;
+    max-width: $form-max-width;
+    margin: 0 auto;
+
+    & > div {
+      margin: 0;
+    }
+  }
+
+  &__paper {
+    width: 100%;
+  }
+
+  &__dialogContent {
+    background-color: var(--bg-color);
   }
 }
